fix: resolve eslint error in AuthContext
