+++ frontend/src/components/Header.tsx
@@ -1,5 +1,5 @@
 import { Link, useNavigate } from "react-router-dom";
-import { useAuth } from "../context/AuthContext";
+import { useAuth } from "../context/useAuth";
 
 export function Header() {
   const { user, logout } = useAuth();
