+++ frontend/src/pages/AdminPage.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import { Link } from "react-router-dom";
-import { useAuth } from "../context/AuthContext";
+import { useAuth } from "../context/useAuth";
 import { createInvite, listInvites } from "../api";
 import type { Invite } from "../types";
