+++ frontend/src/pages/RegisterPage.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { useNavigate, Link } from "react-router-dom";
-import { useAuth } from "../context/AuthContext";
+import { useAuth } from "../context/useAuth";
 
 export function RegisterPage() {
   const [username, setUsername] = useState("");
